add sklearn dep, fitness, rename type to types to not interfere with builtin type
