+++ pyproject.toml
@@ -3,6 +3,7 @@ dependencies = [
   "graphviz>=0.20.3",
   "numpy>=2.2.1",
   "loguru>=0.7.3",
+  "scikit-learn>=1.6.1",
 ]
 description = "Add your description here"
 name = "giraffe"
